RA-1285 Refactored Structure Map configuration in to one file.
