+++ src/Esfa.Vacancy.Register.Api/App_Start/WebApiConfig.cs
@@ -1,5 +1,4 @@
 using System.Web.Http;
-using System.Web.Http.ExceptionHandling;
 using Esfa.Vacancy.Register.Api.App_Start;
 using Newtonsoft.Json;
 
@@ -12,7 +11,8 @@ namespace Esfa.Vacancy.Register.Api
             // Web API configuration and services
 
             // JSON formatters
-            config.Formatters.JsonFormatter.SerializerSettings = new JsonSerializerSettings {
+            config.Formatters.JsonFormatter.SerializerSettings = new JsonSerializerSettings
+            {
                 NullValueHandling = NullValueHandling.Ignore,
                 DateFormatHandling = DateFormatHandling.IsoDateFormat,
                 DateTimeZoneHandling = DateTimeZoneHandling.Utc
@@ -21,7 +21,7 @@ namespace Esfa.Vacancy.Register.Api
             ApiFilterConfig.RegisterFilters(config.Filters);
 
             GlobalConfiguration.Configuration.Formatters.JsonFormatter.SerializerSettings.Converters.Add(new StrictEnumConverter());
-            
+
             // Web API routes
             config.MapHttpAttributeRoutes();
             config.Routes.MapHttpRoute(
@@ -29,9 +29,6 @@ namespace Esfa.Vacancy.Register.Api
                 routeTemplate: "api/{controller}/{id}",
                 defaults: new { id = RouteParameter.Optional }
             );
-
-            var validationBadRequestBuilder = (IValidationBadRequestBuilder)GlobalConfiguration.Configuration.DependencyResolver.GetService(typeof(IValidationBadRequestBuilder));
-            config.Services.Replace(typeof(IExceptionHandler), new VacancyApiExceptionHandler(validationBadRequestBuilder));
         }
     }
 }
